+++ BuildTools/get_git_version.xml
@@ -2,7 +2,7 @@
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
 
   <PropertyGroup>
-    <MajorVersion>2.0</MajorVersion>
+    <MajorVersion>3.0</MajorVersion>
   </PropertyGroup>
 
   <UsingTask TaskName="GitVersion" TaskFactory="CodeTaskFactory" AssemblyFile="$(MSBuildToolsPath)\Microsoft.Build.Tasks.v4.0.dll" >
